PopupModal파일 CommunityWrite로 수정
