--- src/css/PopupMadal.module.css
@@ -1,111 +1,0 @@
-@import url("https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard.min.css");
-
-.popup {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5); 
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 9999; /* 다른 요소 위에 나타나도록 z-index 값을 설정합니다. */
-}
-
-.popupContent {
-    position: relative;
-    background-color: white;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2); 
-    width: 400px;
-    height:100%;
-    padding: 30px;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.titleDiv{
-    position: absolute;
-    top: 30px;
-    left :10px;
-    
-    height:fit-content;
-    
-    display: flex;
-  
-}
-.title{
-    text-align: center;
-    line-height: 60px;
-    font-weight: 800;
-    color: black;
-    font-size: 20px;
-    margin-left: 20px;
-}
-.closeButton {
-    line-height: 50px;
-    text-align: center;
-    cursor: pointer;
-    font-size: 50px;
-    font-weight: 100;
-    color: black;
-}
-.comWriterBox{
-    position: absolute;
-
-    top: 100px;
-    font-size: 20px;
-    
-    color: black;
-
-    width: 90%;
-    height: 70px;
-
-    display: flex;
-    align-items: center;
-    justify-items: center;
-    float: lef;
-    border-bottom: 1.8px solid black;
-    
-    
-}
-.writerProfile{
-    height:45px; width:45px;
-    background-color: #D9D9D9;
-    border-radius: 50px;
-    line-height: 50px;
-    img {
-        width: 50%;
-        height:50%;
-        object-fit: cover; 
-      }
-      
-}
-#writerId{
-    margin-left: 20px;
-}
-
-
-.decoBox{
-    position: absolute;
-    bottom: 50px;
-}
-.questionBox {
-    display: block;
-    width: 400px;
-    height: 400px;
-    
-    padding: 20px 10px;
-    
-    margin: 20px auto;
-    resize: none;
-    border: 1 px solid #aaa;
-    
-    border-radius: 4px;
-    font-size: 16px;
-}
-.btn{
-    float: right;
-}
